Add files via upload
I used a variety of methods to produce a horizontal stacked column bar chart for the five pillars. To construct the graphs, the spi_index_labeled.csv file was inserted into Tableau, and the average score for each pillar was also computed. The five calculated average pillar scores were listed in the columns after the region table in the rows. I chose to use the median as the value for the value aggregation since it more accurately captures the data than the mean, which might be influenced by outliers. I added measure names to the color field to distinguish between the average pillars. In the end, I filtered the data to just display the 2019 results, which were the most recent data accessible.
A horizontal stacked column bar chart is what I utilized as it simplifies the comparison of the median score for each region and the relative weights that each pillar contributed to that score. The pillars are stacked horizontally to make it easy to compare their sizes, and each pillar is identified by its respective color. Also, by showing the chart horizontally, it is simpler to label each bar with the area name without having to change the size of the chart or the text's angle.
According to the information shown in Figure 1, it is clear that averaged pillar 1 performs well overall in 2019 as seen by the high median score it receives across all areas. On the other side, the median score of the averaged pillar 5th  is high exclusively in North America, indicating that there may be unique circumstances or circumstances unique to this area that influence the performance of this pillar. In contrast, as seen by the fact that averaged pillar 5 has the lowest median score across four regions, averaged pillar 4 has the lowest median score across two regions, and averaged pillar 3 has the lowest median score across one region, these pillars perform underwhelmingly in those regions. 
In conclusion, the horizontal stacked column bar chart, with the median as the aggregate value, is a useful tool for showing the scores for the averaged five pillars by region.  By using this style of the chart and restricting the data to 2019, it is achievable to ensure the accuracy and clarity of the data's visual representation.